+++ packages/preview-middleware-client/test/unit/cpe/changes/service.test.ts
@@ -1,10 +1,7 @@
 import * as flexChange from '../../../../src/cpe/changes/flex-change';
 import { ChangeService } from '../../../../src/cpe/changes/service';
-import {
-    changeProperty,
-    deletePropertyChanges,
-    reloadApplication
-} from '@sap-ux-private/control-property-editor-common';
+import { ActionHandler } from '../../../../src/cpe/types';
+import { changeProperty, deletePropertyChanges, numberOfChangesRequiringReloadChanged } from '@sap-ux-private/control-property-editor-common';
 import RuntimeAuthoringMock from 'mock/sap/ui/rta/RuntimeAuthoring';
 import { RTAOptions } from 'sap/ui/rta/RuntimeAuthoring';
 
@@ -14,13 +11,13 @@ describe('SelectionService', () => {
         return Promise.resolve();
     });
     let sendActionMock: jest.Mock;
-    let subscribeMock: jest.Mock;
+    let subscribeMock: jest.Mock<void, [ActionHandler]>;
     const rtaMock = new RuntimeAuthoringMock({} as RTAOptions);
 
     beforeEach(() => {
         rtaMock.attachUndoRedoStackModified = jest.fn() as jest.Mock;
         sendActionMock = jest.fn();
-        subscribeMock = jest.fn();
+        subscribeMock = jest.fn<void, [ActionHandler]>();
         fetchMock.mockClear();
     });
 
@@ -97,7 +94,7 @@ describe('SelectionService', () => {
                     {
                         changeType: 'propertyChange',
                         type: 'saved',
-                        kind: 'valid',
+                        kind: 'property',
                         fileName: 'id_1640106755570_204_propertyChange',
                         controlName: 'Button',
                         controlId:
@@ -107,6 +104,7 @@ describe('SelectionService', () => {
                         timestamp: 1640106817301
                     },
                     {
+                        changeType: 'propertyChange',
                         type: 'saved',
                         kind: 'unknown',
                         controlId:
@@ -117,7 +115,7 @@ describe('SelectionService', () => {
                     {
                         changeType: 'propertyChange',
                         type: 'saved',
-                        kind: 'valid',
+                        kind: 'property',
                         fileName: 'id_1640106755570_203_propertyChange',
                         controlName: 'Button',
                         controlId:
@@ -179,12 +177,13 @@ describe('SelectionService', () => {
                         type: 'saved',
                         kind: 'unknown',
                         fileName: 'unknown',
+                        changeType: 'addXML',
                         timestamp: 1640106877301,
                         controlId: 'SEPMRA_C_PD_Product--template::ListReport::TableToolbar'
                     },
                     {
                         type: 'saved',
-                        kind: 'valid',
+                        kind: 'property',
                         changeType: 'propertyChange',
                         fileName: 'id_1640106755570_204_propertyChange',
                         controlName: 'Button',
@@ -271,6 +270,7 @@ describe('SelectionService', () => {
                         isActive: true,
                         controlName: 'SimpleForm',
                         fileName: 'testFileName',
+                        kind: 'unknown',
                         type: 'pending'
                     },
                     {
@@ -279,6 +279,7 @@ describe('SelectionService', () => {
                         isActive: true,
                         controlName: 'SimpleForm',
                         fileName: 'testFileName',
+                        kind: 'unknown',
                         type: 'pending'
                     }
                 ]
@@ -301,7 +302,8 @@ describe('SelectionService', () => {
                 }),
                 getPreparedChange: jest.fn().mockReturnValue({
                     getDefinition: jest.fn().mockReturnValue({
-                        changeType: 'page'
+                        changeType: 'page',
+                        fileName: 'fileName'
                     })
                 })
             };
@@ -334,6 +336,8 @@ describe('SelectionService', () => {
                         controlId: '_ST_SmartVariantManagement',
                         isActive: true,
                         controlName: 'SimpleForm',
+                        kind: 'unknown',
+                        fileName: 'fileName',
                         type: 'pending'
                     },
                     {
@@ -341,6 +345,8 @@ describe('SelectionService', () => {
                         controlId: '_ST_SmartVariantManagement',
                         isActive: true,
                         controlName: 'SimpleForm',
+                        kind: 'unknown',
+                        fileName: 'fileName',
                         type: 'pending'
                     }
                 ]
@@ -472,6 +478,7 @@ describe('SelectionService', () => {
                         controlName: 'Button',
                         fileName: 'testFileName',
                         type: 'pending',
+                        kind: 'property',
                         value: 'abc'
                     },
                     {
@@ -482,6 +489,7 @@ describe('SelectionService', () => {
                         controlName: 'Button',
                         fileName: 'testFileName',
                         type: 'pending',
+                        kind: 'property',
                         value: '{i18n>DELETE}'
                     },
                     {
@@ -490,6 +498,7 @@ describe('SelectionService', () => {
                         controlName: 'ExtensionPoint1',
                         fileName: 'testFileName',
                         isActive: true,
+                        kind: 'unknown',
                         type: 'pending'
                     }
                 ]
@@ -527,6 +536,89 @@ describe('SelectionService', () => {
             controlName: 'button',
             propertyName: 'text',
             value: 'abc'
+        });
+    });
+
+    test('manifest change', async () => {
+        fetchMock.mockResolvedValue({ json: () => Promise.resolve({}) });
+        function createCommand(
+            properties: Map<string, any>,
+            toggle = false
+        ): {
+            getProperty: (name: string) => any;
+            getElement: () => any;
+            getSelector: () => any;
+            getChangeType: () => string;
+            getParent: () => any;
+            getPreparedChange: () => { getDefinition: () => { fileName: string } };
+        } {
+            const cache = new Map(properties);
+            return {
+                getProperty: (name: string): any => {
+                    return cache.get(name);
+                },
+                getElement: jest.fn().mockReturnValue({
+                    getMetadata: jest.fn().mockReturnValue({ getName: jest.fn().mockReturnValue('sap.m.Button') })
+                }),
+                getSelector: jest.fn().mockReturnValue({
+                    id: !toggle ? 'ListReport.view.ListReport::SEPMRA_C_PD_Product--app.my-test-button' : undefined,
+                    name: 'ExtensionPoint1'
+                }),
+                getChangeType: (): any => {
+                    return cache.get('changeType');
+                },
+                getParent: jest.fn().mockReturnValue({
+                    getElement: jest.fn().mockReturnValue({
+                        getId: () => 'ListReport.view.ListReport::SEPMRA_C_PD_Product--app.my-test-button'
+                    })
+                }),
+                getPreparedChange: (): { getDefinition: () => { fileName: string } } => {
+                    return { getDefinition: () => ({ fileName: 'testFileName' }) };
+                }
+            };
+        }
+        const commands = [
+            createCommand(
+                new Map<string, any>([
+                    ['selector', { id: 'control1' }],
+                    ['changeType', 'appdescr_fe_changePageConfiguration'],
+                    ['propertyName', 'text'],
+                    ['newValue', 'abc']
+                ])
+            ),
+        ];
+        rtaMock.getCommandStack.mockReturnValue({
+            getCommands: jest.fn().mockReturnValue(commands),
+            getAllExecutedCommands: jest.fn().mockReturnValue(commands)
+        });
+        const service = new ChangeService(
+            { rta: rtaMock } as any,
+            {
+                applyControlPropertyChange: jest.fn()
+            } as any
+        );
+
+        await service.init(sendActionMock, subscribeMock);
+
+        await (rtaMock.attachUndoRedoStackModified as jest.Mock).mock.calls[0][0]();
+        expect(sendActionMock).toHaveBeenCalledTimes(5);
+        expect(sendActionMock).toHaveBeenNthCalledWith(2, numberOfChangesRequiringReloadChanged(1))
+        expect(sendActionMock).toHaveBeenNthCalledWith(3, {
+            type: '[ext] change-stack-modified',
+            payload: {
+                saved: [],
+                pending: [
+                    {
+                        changeType: 'appdescr_fe_changePageConfiguration',
+                        controlId: 'ListReport.view.ListReport::SEPMRA_C_PD_Product--app.my-test-button',
+                        isActive: true,
+                        controlName: 'Button',
+                        fileName: 'testFileName',
+                        type: 'pending',
+                        kind: 'unknown',
+                    },
+                ]
+            }
         });
     });
 
@@ -572,55 +664,6 @@ describe('SelectionService', () => {
             body: '{"fileName":"id_1640106755570_203_propertyChange"}',
             headers: { 'Content-Type': 'application/json' },
             method: 'DELETE'
-        });
-    });
-
-    test('reload application', async () => {
-        jest.spyOn(Date, 'now').mockReturnValueOnce(123);
-        fetchMock.mockResolvedValue({
-            json: () => Promise.resolve(undefined)
-        });
-
-        const service = new ChangeService(
-            { rta: rtaMock } as any,
-            {
-                applyControlPropertyChange: jest.fn()
-            } as any
-        );
-
-        await service.init(sendActionMock, subscribeMock);
-        await subscribeMock.mock.calls[0][0](reloadApplication());
-        expect(rtaMock.stop).toHaveBeenNthCalledWith(1, false, true);
-    });
-
-    test('attach stop callback check', async () => {
-        jest.spyOn(Date, 'now').mockReturnValueOnce(123);
-        fetchMock.mockResolvedValue({
-            json: () => Promise.resolve(undefined)
-        });
-
-        const service = new ChangeService(
-            { rta: rtaMock } as any,
-            {
-                applyControlPropertyChange: jest.fn()
-            } as any
-        );
-
-        rtaMock.attachStop.mockClear();
-        const reloadSpy = jest.fn();
-        const location = window.location;
-        Object.defineProperty(window, 'location', {
-            value: {
-                reload: reloadSpy
-            }
-        });
-        await service.init(sendActionMock, subscribeMock);
-        expect(rtaMock.attachStop).toBeCalledTimes(1);
-
-        rtaMock.attachStop.mock.calls[0][0]();
-        expect(reloadSpy).toHaveBeenCalled();
-        Object.defineProperty(window, 'location', {
-            value: location
         });
     });
 });