+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/app-config-writer
 
+## 0.4.34
+
+### Patch Changes
+
+-   Updated dependencies [070182d]
+    -   @sap-ux/project-access@1.27.3
+    -   @sap-ux/axios-extension@1.16.6
+
 ## 0.4.33
 
 ### Patch Changes