+++ packages/launch-config/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/launch-config
 
+## 0.5.0
+
+### Minor Changes
+
+-   463f482: Reverted the use of Node.js `fs` modules and replaced them with `mem-fs` for writing launch config files & Removed `writeApplicationInfoSettings()` from `@sap-ux/launch-config`
+    Refactoring create launch config functionalities.
+
+## 0.4.4
+
+### Patch Changes
+
+-   Updated dependencies [070182d]
+    -   @sap-ux/project-access@1.27.3
+    -   @sap-ux/odata-service-inquirer@0.5.40
+
 ## 0.4.3
 
 ### Patch Changes